Render of svg code segment is done.
This can be verified from "cargo run" at the top level of the directory
