+++ src/bin/render.rs
@@ -1,0 +1,50 @@
+extern crate handlebars;
+
+use std::collections::BTreeMap;
+use handlebars::Handlebars;
+use std::fs::File;
+use std::path::Path;
+use std::io::{self, BufReader, BufRead};
+
+fn render_source_code(path: &str) -> String {
+    
+    /* Template creation */
+    let mut handlebars = Handlebars::new();
+    let LINE_TEMPLATE = "                <text class=\"code\" x=\"{{X_VAL}}\" y=\"{{Y_VAL}}\"> {{LINE}} </text>\n";
+    // register the template. The template string will be verified and compiled.
+    assert!(handlebars.register_template_string("code_line_template", LINE_TEMPLATE).is_ok());
+
+    /* Render the code segment of the svg to a String */
+    let x = 0;
+    let mut y = 80;
+    let mut output = String::from("        <g id=\"code\">\n");
+    if let Ok(lines) = read_lines(path.to_owned() + "/annotated_source.rs") {
+        // Consumes the iterator, returns an (Optional) String
+        for line in lines {
+            if let Ok(line_string) = line {
+                let mut data = BTreeMap::new();
+                data.insert("X_VAL".to_string(), x.to_string());
+                data.insert("Y_VAL".to_string(), y.to_string());
+                data.insert("LINE".to_string(), line_string);
+                output.push_str(&handlebars.render("code_line_template", &data).unwrap());
+                y = y + 20;
+            }
+        }
+    }
+    output.push_str("        </g>\n");
+    output
+}
+
+fn main() {
+    let path = "example/move_with_no_borrow";
+    let rendered_string = render_source_code(&path);
+    println!("{}", rendered_string);
+}
+
+// File reader helper function.
+// Returns: a result object that we can match to OK(), and then use .lines() on.
+fn read_lines<P>(filename: P) -> io::Result<io::Lines<io::BufReader<File>>>
+where P: AsRef<Path>, {
+    let file = File::open(filename)?;
+    Ok(BufReader::new(file).lines())
+}
